+++ DataService/src/main/java/com/crafters/DataService/dtos/ItemResponse.java
@@ -18,7 +18,7 @@ public class ItemResponse {
     private String rowType;
     private String collectionName;
     private Map<String, Object> attributes;
-    private Map<String, Integer> yearValue;
+    private Map<String, Double> yearValue;
 
     public ItemResponse(Item item) {
         this.id = item.getId();
